Moved multi day av tests into own file
