+++ autoscheduler/frontend/src/utils/timeUtil.ts
@@ -15,3 +15,21 @@ export function formatTime(h: number, m: number, use24Hour = false, padZeroes = 
 
   return `${padZeroes ? padZero.format(formattedHours) : formattedHours}:${padZero.format(m)}`;
 }
+
+/**
+ * Creates fake mouse events at a certain time. For use in testing.
+ * @param h hours of the desired time, in 24-hr format
+ * @param m minutes of the desired time
+ * @param offset top of the calendar day in which the event is fired
+ * @param clientHeight height of the calendar day
+ */
+export const timeToEvent = (h: number, m: number, offset = 0, clientHeight = 1000): {} => {
+  const minsPastStart = h * 60 + m - FIRST_HOUR * 60;
+  const minsPerDay = (LAST_HOUR - FIRST_HOUR) * 60;
+
+  return {
+    button: 0,
+    clientY: offset + minsPastStart / minsPerDay * clientHeight,
+    clientX: 100,
+  };
+};
